feat: complete refactor to Cursor Enhancer with audio removal and Linux focus
## Major Changes

### Project Rebranding
- Renamed from "Review Gate V2" to "Cursor Enhancer"
- Updated all file names, class names, and references
- Created new rule file: CursorEnhancerV2.mdc
- Updated package.json with new branding

### Audio Functionality Removal
- Removed all speech-to-text processing (Faster-Whisper)
- Eliminated audio UI components and recording logic
- Removed SoX dependency from installation scripts
- Simplified to text + image input only
- Reduced requirements.txt dependencies

### Modular Architecture Refactoring
- Implemented Single Responsibility Principle
- Created modular src/ structure with focused components:
  - src/config/ - Centralized configuration constants
  - src/services/ - Business logic and service classes
  - src/managers/ - Response and trigger file managers
  - src/protocol/ - MCP protocol handling
  - src/utils/ - Shared utilities and logging
- Refactored monolithic files into 246-line focused modules
- Applied DRY principles to eliminate code duplication

### Platform Simplification
- Removed Windows and macOS support completely
- Linux Ubuntu/Debian only with apt-get integration
- Simplified installation scripts and dependencies
- Updated documentation to reflect Linux-only focus

### File Structure Updates
- Main server: review_gate_v2_mcp.py → cursor_enhancer_mcp.py
- Tool name: review_gate_chat → cursor_enhancer_chat
- File patterns: review_gate_trigger → cursor_enhancer_trigger
- MCP config: review-gate-v2 → cursor-enhancer

### Extension Updates
- Modular popup-manager.js and file-watcher.js
- Removed audio-handler.js completely
- Updated UI to remove microphone button
- Streamlined interface for text + images

### Technical Improvements
- Faster startup (no ML model loading)
- Reduced complexity and failure points
- Cleaner dependency management
- Better error handling and logging
- Improved file-based communication protocol
