+++ src/__init__.py
@@ -1,0 +1,15 @@
+"""
+Review Gate V2 - Modular Python Package
+
+This package contains the refactored components for Review Gate V2:
+- config: Configuration constants and settings
+- utils: Shared utility functions for file operations and logging
+- protocol: MCP protocol handling
+- services: Business logic and tool execution
+- managers: Response and trigger management
+
+All components follow Single Responsibility Principle and eliminate DRY violations.
+"""
+
+__version__ = "2.0.0"
+__author__ = "Lakshman Turlapati"
