+++ src/config/__init__.py
@@ -1,0 +1,5 @@
+"""Configuration module for Review Gate V2."""
+
+from .constants import FilePatterns, TimeoutConfig
+
+__all__ = ["TimeoutConfig", "FilePatterns"]
