+++ src/managers/__init__.py
@@ -1,0 +1,6 @@
+"""Manager modules for Review Gate V2."""
+
+from .response_manager import ResponseManager
+from .trigger_manager import TriggerManager
+
+__all__ = ["ResponseManager", "TriggerManager"]
