+++ src/protocol/__init__.py
@@ -1,0 +1,5 @@
+"""MCP Protocol handling module for Review Gate V2."""
+
+from .mcp_handler import McpProtocolHandler
+
+__all__ = ["McpProtocolHandler"]
