+++ src/services/__init__.py
@@ -1,0 +1,6 @@
+"""Service modules for Cursor Enhancer."""
+
+from .cursor_enhancer_service import CursorEnhancerService
+from .tool_executor import ToolExecutor
+
+__all__ = ["ToolExecutor", "CursorEnhancerService"]
