+++ src/services/cursor_enhancer_service.py
@@ -1,0 +1,21 @@
+"""
+Cursor Enhancer service module containing business logic for Cursor Enhancer operations.
+"""
+
+import logging
+from typing import Any
+
+
+class CursorEnhancerService:
+    """Service class for Cursor Enhancer business logic operations"""
+
+    def __init__(self):
+        self.logger = logging.getLogger(__name__)
+
+    def validate_tool_arguments(self, tool_name: str, arguments: dict[str, Any]) -> bool:
+        """Validate tool arguments"""
+        if tool_name == "cursor_enhancer_chat":
+            # Basic validation for cursor_enhancer_chat
+            return True  # Currently accepts all arguments
+
+        return False
